requirements.txt
numpy ~= 1.26.4
scipy ~= 1.13.1
matplotlib ~= 3.9.2
PyQt5 ~= 5.15.10
pillow ~= 10.4.0